API 20160730 01 bruno - [bug fix] could not create a project because of parent didn't exists - [bug fix] sometime could not create a symlink on production server

+++ libs/Folders.php
@@ -93,10 +93,12 @@ class Folders {
 	}
 
 	public function createSymlink($target, $link){
-		if($this->createPath($link)){
-			if(rmdir($link)){
-				if(symlink($target, $link)){
-					return $this->setPath($link);
+		if(!is_dir($link)){
+			if($this->createPath($link)){
+				if(rmdir($link)){
+					if(symlink($target, $link)){
+						return $this->setPath($link);
+					}
 				}
 			}
 		}
